@@ -6,7 +6,7 @@
     "postinstall": "ng build --aot --prod --build-optimizer=false",
     "ng": "ng",
     "start": "node server.js",
-    "hmr": "ng serve --hmr -e=hmr",
+    "hmr": "ng serve --hmr --e=hmr",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
@@ -17,61 +17,62 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "5.0.0-beta.15",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@ngrx/effects": "^4.0.0",
-    "@ngrx/router-store": "^4.0.0",
-    "@ngrx/store": "^4.0.0",
-    "@ngrx/store-devtools": "^4.0.0",
-    "ag-grid": "^18.0.1",
-    "ag-grid-angular": "^18.0.1",
-    "angularfire2": "^5.0.0-rc.7.0-next",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^6.1.7",
+    "@angular/cdk": "^6.4.7",
+    "@angular/common": "^6.1.7",
+    "@angular/compiler": "^6.1.7",
+    "@angular/compiler-cli": "^6.1.7",
+    "@angular/core": "^6.1.7",
+    "@angular/fire": "^5.0.0",
+    "@angular/flex-layout": "^6.0.0-beta.18",
+    "@angular/forms": "^6.1.7",
+    "@angular/http": "^6.1.7",
+    "@angular/material": "^6.4.7",
+    "@angular/platform-browser": "^6.1.7",
+    "@angular/platform-browser-dynamic": "^6.1.7",
+    "@angular/router": "^6.1.7",
+    "@ngrx/effects": "6.1.0",
+    "@ngrx/router-store": "6.1.0",
+    "@ngrx/store": "6.1.0",
+    "@ngrx/store-devtools": "6.1.0",
+    "ag-grid": "^18.1.2",
+    "ag-grid-angular": "^18.1.0",
+    "core-js": "^2.5.7",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
-    "firebase": "^4.13.1",
+    "firebase": "^5.4.2",
     "hammerjs": "^2.0.8",
     "jwt-decode": "^2.2.0",
     "moment": "^2.22.2",
     "ng2-search-filter": "^0.4.7",
-    "ngrx-actions": "^2.1.1",
-    "ngrx-store-freeze": "^0.2.0",
+    "ngrx-actions": "^4.0.0",
+    "ngrx-store-freeze": "^0.2.4",
     "ngx-order-pipe": "^2.0.1",
-    "rxjs": "^5.5.6",
-    "typescript": "~2.5.3",
-    "yargs": "^11.0.0",
+    "rxjs": "^6.3.2",
+    "typescript": "~2.9.2",
+    "yargs": "^11.1.0",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.3",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/cli": "^6.2.1",
+    "@angular/language-service": "^6.1.7",
     "@angularclass/hmr": "^2.1.3",
-    "@types/jasmine": "~2.8.3",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "enhanced-resolve": "^4.0.0",
+    "@types/node": "^6.0.117",
+    "codelyzer": "^4.4.4",
+    "enhanced-resolve": "^4.1.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.5",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "node-sass": "^4.9.3",
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "@angular-devkit/build-angular": "~0.6.3"
+    "tslint": "~5.9.1"
   }
 }
