@@ -3,7 +3,7 @@ import {AuthService} from '../../services/auth.service';
 import {IAppState} from '../../store/store';
 import {Store} from '@ngrx/store';
 import {getLastUpdated, getNumberOne} from '../../store/participanttable/participanttable.reducer';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {IParticipanttable} from '../../models/participanttable.model';
 import {getTour, isRegistrationOpen} from '../../store/tour/tour.reducer';
 import {ITour} from '../../models/tour.model';
