@@ -1,7 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Router} from '@angular/router';
 import {AuthService} from './auth.service';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
+import {of} from 'rxjs';
 
 @Injectable()
 export class AdminGuardService {
@@ -20,7 +21,7 @@ export class AdminGuardService {
       }
     }).catch(() => {
       this.router.navigate(['/login']);
-      return Observable.of(false);
+      return of(false);
     });
   }
 }
