@@ -1,8 +1,10 @@
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {AngularFireAuth} from 'angularfire2/auth';
+import {Observable} from 'rxjs';
 import {Router} from '@angular/router';
-import * as firebase from 'firebase';
+import * as firebase from 'firebase/app';
+import 'firebase/auth';
+// import * as firebase from 'firebase';
+import {AngularFireAuth} from '@angular/fire/auth';
 
 @Injectable()
 export class AuthService {
