@@ -1,8 +1,9 @@
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {environment} from '../../environments/environment';
 import {IParticipant} from '../models/participant.model';
 import {HttpClient} from '@angular/common/http';
+import {map} from 'rxjs/operators';
 
 @Injectable()
 export class ParticipantService {
@@ -21,6 +22,6 @@ export class ParticipantService {
 
   postParticipant(body: IParticipant): Observable<IParticipant> {
     return this.http.post<IParticipant>(`${environment.apiBaseUrl}/participants`, body)
-      .map(res => <IParticipant>res);
+      .pipe(map(res => <IParticipant>res));
   }
 }
