@@ -1,8 +1,9 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {IStageClassification, ITourClassification} from '../models/etappe.model';
 import {environment} from '../../environments/environment';
+import {map} from 'rxjs/operators';
 
 @Injectable()
 export class ClassificationsService {
@@ -12,7 +13,7 @@ export class ClassificationsService {
 
   saveStageclassifications(body: IStageClassification[]): Observable<any> {
     return this.http.post<IStageClassification[]>(`${environment.apiBaseUrl}/stageclassifications`, body)
-      .map(res => <IStageClassification[]>res);
+      .pipe(map(res => <IStageClassification[]>res));
   }
 
   getStageClassifications(etappeId: string): Observable<IStageClassification[]> {
@@ -21,7 +22,7 @@ export class ClassificationsService {
 
   saveTourclassifications(body: ITourClassification[]): Observable<any> {
     return this.http.post<ITourClassification[]>(`${environment.apiBaseUrl}/tourclassifications`, body)
-      .map(res => <ITourClassification[]>res);
+      .pipe(map(res => <ITourClassification[]>res));
   }
 
   getTourClassifications(tourId: string): Observable<ITourClassification[]> {
@@ -30,7 +31,7 @@ export class ClassificationsService {
 
   saveYouthclassifications(body: ITourClassification[]): Observable<any> {
     return this.http.post<ITourClassification[]>(`${environment.apiBaseUrl}/youthclassifications`, body)
-      .map(res => <ITourClassification[]>res);
+      .pipe(map(res => <ITourClassification[]>res));
   }
 
   getYouthClassifications(tourId: string): Observable<ITourClassification[]> {
@@ -39,14 +40,16 @@ export class ClassificationsService {
 
   saveMountainclassifications(body: ITourClassification[]): Observable<any> {
     return this.http.post<ITourClassification[]>(`${environment.apiBaseUrl}/mountainclassifications`, body)
-      .map(res => <ITourClassification[]>res);
+      .pipe(map(res => <ITourClassification[]>res));
   }
 
   getMountainClassifications(tourId: string): Observable<ITourClassification[]> {
     return this.http.get<ITourClassification[]>(`${environment.apiBaseUrl}/mountainclassifications/${tourId}`);
-  } savePointsclassifications(body: ITourClassification[]): Observable<any> {
+  }
+
+  savePointsclassifications(body: ITourClassification[]): Observable<any> {
     return this.http.post<ITourClassification[]>(`${environment.apiBaseUrl}/pointsclassifications`, body)
-      .map(res => <ITourClassification[]>res);
+      .pipe(map(res => <ITourClassification[]>res));
   }
 
   getPointsClassifications(tourId: string): Observable<ITourClassification[]> {
