@@ -1,12 +1,10 @@
 import {Injectable} from '@angular/core';
 import {Actions, Effect} from '@ngrx/effects';
 import * as rider from './rider.actions';
-import {Observable} from 'rxjs/Observable';
 import {RiderService} from '../../services/rider.service';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/catch';
+import {catchError, switchMap} from 'rxjs/operators';
+import {of} from 'rxjs/internal/observable/of';
+
 
 @Injectable()
 export class RiderEffects {
@@ -17,12 +15,11 @@ export class RiderEffects {
   @Effect()
   fetchRider$ = this.actions$
     .ofType<rider.FetchRiders>(rider.FETCH_RIDERS)
-    .switchMap(action => {
+    .pipe(switchMap(action => {
       return this.riderService
         .getRiders()
-        .switchMap(riderResponse =>
-          Observable.of(new rider.FetchRidersSuccess(riderResponse))
-        )
-        .catch(err => Observable.of(new rider.FetchRidersFailure(err)));
-    });
+        .pipe(switchMap(riderResponse =>
+            of(new rider.FetchRidersSuccess(riderResponse))),
+          catchError(err => of(new rider.FetchRidersFailure(err))));
+    }));
 }
