@@ -1,12 +1,11 @@
 import {Injectable} from '@angular/core';
 import {Actions, Effect} from '@ngrx/effects';
 import * as etappe from './etappe.actions';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/catch';
+
+
 import {EtappeService} from '../../services/etappe.service';
+import {catchError, switchMap} from 'rxjs/operators';
+import {of} from 'rxjs/internal/observable/of';
 
 @Injectable()
 export class EtappeEffects {
@@ -17,12 +16,11 @@ export class EtappeEffects {
   @Effect()
   fetchEtappeList$ = this.actions$
     .ofType<etappe.FetchEtappeList>(etappe.FETCH_ETAPPELIST)
-    .switchMap(action => {
+    .pipe(switchMap(action => {
       return this.etappeService
         .getEtappes(action.payload)
-        .switchMap(etappeResponse =>
-          Observable.of(new etappe.FetchEtappeListSuccess(etappeResponse))
-        )
-        .catch(err => Observable.of(new etappe.FetchEtappeListFailure(err)));
-    });
+        .pipe(switchMap(etappeResponse =>
+            of(new etappe.FetchEtappeListSuccess(etappeResponse))),
+          catchError(err => of(new etappe.FetchEtappeListFailure(err))));
+    }));
 }
