@@ -1,12 +1,10 @@
 import {Injectable} from '@angular/core';
 import {Actions, Effect} from '@ngrx/effects';
 import * as team from './team.actions';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/catch';
+
 import {TeamService} from '../../services/teams.service';
+import {catchError, switchMap} from 'rxjs/operators';
+import {of} from 'rxjs/internal/observable/of';
 
 @Injectable()
 export class TeamEffects {
@@ -17,11 +15,11 @@ export class TeamEffects {
   @Effect()
   fetchRider$ = this.actions$
     .ofType<team.FetchTeams>(team.FETCH_TEAMS)
-    .switchMap(action => {
+    .pipe(switchMap(action => {
       return this.teamService.getTeams()
-        .switchMap(teamResponse =>
-          Observable.of(new team.FetchTeamsSuccess(teamResponse))
-        )
-        .catch(err => Observable.of(new team.FetchTeamsFailure(err)));
-    });
+        .pipe(switchMap(teamResponse =>
+            of(new team.FetchTeamsSuccess(teamResponse))
+          ),
+          catchError(err => of(new team.FetchTeamsFailure(err))));
+    }));
 }
